1.9.2 Update

+++ src/com/loohp/interactionvisualizer/Manager/LightManager.java
@@ -1,7 +1,6 @@
 package com.loohp.interactionvisualizer.Manager;
 
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.Map.Entry;
 import java.util.Queue;
@@ -32,29 +31,16 @@ public class LightManager {
 		deletequeue.add(location);
 	}
 	
-	public static int gc() {
-		return new BukkitRunnable() {
-			public void run() {
-				Iterator<Entry<Location, Integer>> itr = lights.entrySet().iterator();
-				while (itr.hasNext()) {
-					Entry<Location, Integer> entry = itr.next();
-					if (deletequeue.contains(entry.getKey())) {
-						deletequeue.removeIf(each -> each.equals(entry.getKey()));
-						itr.remove();
-					}
-				}
-			}
-		}.runTaskTimerAsynchronously(InteractionVisualizer.plugin, 0, 3).getTaskId();
-	}
-	
 	public static int run() {
 		return new BukkitRunnable() {
 			public void run() {
 				Set<Location> locations = new HashSet<Location>();
 				while (!deletequeue.isEmpty()) {
 					Location location = deletequeue.poll();
-					LightAPI.deleteLight(location, LightType.BLOCK, false);
-					locations.add(location);
+					if (location != null) {
+						LightAPI.deleteLight(location, LightType.BLOCK, false);
+						locations.add(location);
+					}
 				}
 				for (Entry<Location, Integer> entry : lights.entrySet()) {
 					Location location = entry.getKey();
@@ -62,6 +48,7 @@ public class LightManager {
 					LightAPI.createLight(location, LightType.BLOCK, lightlevel, false);
 					locations.add(location);
 				}
+				lights.clear();
 				Queue<ChunkInfo> infos = new LinkedList<ChunkInfo>();
 				for (Location location : locations) {
 					infos.addAll(LightAPI.collectChunks(location, LightType.BLOCK, 15));
